Mastern

+++ share/pulseaudio-equalizer/pulseaudio-equalizer.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 # PulseAudio Equalizer (PyGtk Interface)
-# version: 2021.11
+# version: 2022.08
 #
 # Maintainer: Luis Armando Medina Avitia <lamedina AT gmail DOT com>
 #
@@ -17,6 +17,8 @@ gi.check_version('3.30')
 gi.require_version('Gtk', '3.0')
 from gi.repository import Gtk, Gio, GLib
 import os, sys
+import shutil
+import subprocess
 
 if 'PULSE_CONFIG_PATH' in os.environ:
     CONFIG_DIR = os.getenv('PULSE_CONFIG_PATH')
@@ -28,6 +30,7 @@ CONFIG_FILE = os.path.join(CONFIG_DIR, 'equalizerrc')
 PRESETS_FILE = os.path.join(CONFIG_DIR, 'equalizerrc.availablepresets')
 USER_PRESET_DIR = os.path.join(CONFIG_DIR, 'presets')
 SYSTEM_PRESET_DIR = os.path.join('/usr/share/pulseaudio-equalizer', 'presets')
+PLUGIN_PRESENT = False
 if os.environ.get('GNOME_DESKTOP_SESSION_ID'):
     DESKTOP_ENVIRONMENT = "Gnome"
 else:
@@ -35,9 +38,12 @@ else:
 
 if not os.path.exists('/usr/lib/ladspa/mbeq_1197.so'):
     print("Warning: Missing pluseaudio plug-in. Please install the swh-plugins package. ")
+else:
+    PLUGIN_PRESENT = True
 
 print(CONFIG_DIR)
 print(SYSTEM_PRESET_DIR)
+print(DESKTOP_ENVIRONMENT)
 
 from urllib.request import url2pathname, urlparse, unquote
 from gi.repository import GdkPixbuf, Gdk
@@ -48,7 +54,8 @@ eqconfig2 = configdir + '/equalizerrc.test'
 eqpresets = eqconfig + '.availablepresets'
 presetdir1 = configdir + '/presets'
 presetdir2 = '/usr/share/pulseaudio-equalizer/presets'
-pulseaudio_scrip =  "../bin/pulseaudio-equalizer" #"pulseaudio-equalizer"
+pulseaudio_scrip =  "/home/lm/Dev/pulseaudio-equalizer-ladspa/bin/pulseaudio-equalizer"
+#pulseaudio_scrip =  "/usr/bin/pulseaudio-equalizer"
 
 TARGET_TYPE_URI_LIST = 80
 
@@ -197,8 +204,8 @@ def FormatLabels(x):
     if len(c) < 2 and len(suffix) == 3:
         whitespace1 = '  '
 
-#@Gtk.Template(filename='../share/pulseaudio-equalizer/equalizer.ui')
-@Gtk.Template(filename='/usr/share/pulseaudio-equalizer/equalizer.ui')
+@Gtk.Template(filename='../share/pulseaudio-equalizer/equalizer.ui')
+#@Gtk.Template(filename='/usr/share/pulseaudio-equalizer/equalizer.ui')
 class Equalizer(Gtk.ApplicationWindow):
     __gtype_name__= "Equalizer"
 
@@ -218,9 +225,14 @@ class Equalizer(Gtk.ApplicationWindow):
     menustd: Gtk.MenuBar = Gtk.Template.Child()
     actionbar: Gtk.ActionBar = Gtk.Template.Child()
 
+    dialog_message: Gtk.Label = Gtk.Template.Child('message')
+
     window_about = Gtk.Template.Child('About')
     window_title: Gtk.Label = Gtk.Template.Child('title')
     window_save = Gtk.Template.Child('Save_dialog')
+
+    window_dialog = Gtk.Template.Child('Message_dialog')
+
 
     def on_scale(self, widget, y):
         global ladspa_controls
@@ -504,31 +516,31 @@ class Equalizer(Gtk.ApplicationWindow):
     @Gtk.Template.Callback()
     def on_delete(self, widget):
         global preset
-        os.remove(os.path.join(USER_PRESET_DIR, preset + '.preset'))
+        if os.path.exists(os.path.join(USER_PRESET_DIR, preset + '.preset')):
+            os.remove(os.path.join(USER_PRESET_DIR, preset + '.preset'))
 
-        self.presetsbox.get_child().set_text('')
-        self.presetsbox1.get_child().set_text('')
+            self.presetsbox.get_child().set_text('')
+            self.presetsbox1.get_child().set_text('')
 
-        # Clear preset list from ComboBox
-        self.presetsbox.remove_all()
-        self.presetsbox1.remove_all()
+            # Clear preset list from ComboBox
+            self.presetsbox.remove_all()
+            self.presetsbox1.remove_all()
+           
+            # Apply settings (which will save new preset as default)
+            ApplySettings()
 
-        # Refresh (and therefore, sort) preset list
-        GetSettings()
+            # Refresh (and therefore, sort) preset list
+            os.system(f'{pulseaudio_scrip} interface.getsettings')
+            GetSettings()
 
-        # Repopulate preset list into ComboBox
-        for i in range(len(rawpresets)):
-            self.presetsbox.append_text(rawpresets[i])
-            self.presetsbox1.append_text(rawpresets[i])
+            # Repopulate preset list into ComboBox
+            for i in range(len(rawpresets)):
+                self.presetsbox.append_text(rawpresets[i])
+                self.presetsbox1.append_text(rawpresets[i])
 
-        preset = ''
-        
-        # Apply settings (which will save new preset as default)
-        ApplySettings()
-
-        # Refresh (and therefore, sort) preset list
-        os.system(f'{pulseaudio_scrip} interface.getsettings')
-        GetSettings()
+            preset = ''
+        else:
+            pass # show dialog window "user preset does not exist."
 
     @Gtk.Template.Callback()
     def on_about_activate(self, widget):
@@ -738,7 +750,38 @@ class Equalizer(Gtk.ApplicationWindow):
             f.close()
 
         dialog.destroy()
-    
+
+    @Gtk.Template.Callback()
+    def on_exportallpresets(self, widget):
+        global preset
+        global presets
+        global presetdir1
+
+        print('on_exportallpresets')
+        dialog = Gtk.FileChooserDialog(title='Export All Presets...',
+                                       parent=None, 
+                                       action=Gtk.FileChooserAction.SELECT_FOLDER)
+        dialog.add_buttons(Gtk.STOCK_CANCEL,Gtk.ResponseType.CANCEL)
+        dialog.add_buttons(Gtk.STOCK_OK, Gtk.ResponseType.OK)
+        dialog.set_default_response(Gtk.ResponseType.OK)
+
+        xdg_bin = shutil.which('xdg-user-dir')
+        process = subprocess.run([xdg_bin, 'DOWNLOAD'], stdout=subprocess.PIPE)
+        download_path = process.stdout.strip().decode()
+
+        dialog.set_current_folder(download_path)
+        dialog.show()
+
+        response = dialog.run()
+        print(response)
+
+        if response == Gtk.ResponseType.OK:
+            export_path = dialog.get_filename()
+            print(export_path)
+            shutil.copytree(presetdir1, export_path, symlinks=False, ignore=None, ignore_dangling_symlinks=False, dirs_exist_ok=True)
+
+        dialog.destroy()
+
     def on_drag_data_received(self, widget, context, x, y, selection, target_type, timestamp):
         global preset
         global presets
@@ -874,6 +917,10 @@ class Equalizer(Gtk.ApplicationWindow):
         action = Gio.SimpleAction.new('on_exportpreset', None)
         action.connect('activate', self.on_exportpreset)
         self.add_action(action)
+
+        action = Gio.SimpleAction.new('on_exportallpresets', None)
+        action.connect('activate', self.on_exportallpresets)
+        self.add_action(action)
         
         action = Gio.SimpleAction.new('on_save', None)
         action.connect('activate', self.on_save)
@@ -917,6 +964,14 @@ class Equalizer(Gtk.ApplicationWindow):
         self.add_action(action)
 
         self.show()
+
+        if not PLUGIN_PRESENT :
+            # when swh plug-ins are not installed a warning message is displayed
+            self.dialog_message.title = "Warning"
+            self.dialog_message.set_text("Warning: Missing pluseaudio plug-in. Please install the swh-plugins package.")
+            result = self.window_dialog.run()
+            if result == -3 :
+                self.window_dialog.hide()
         
 
 class FrequencyLabel(Gtk.Label):
